tuna: Set TARGET_SETS_FSTAB to true.
This changes the updater-script to have mount automatically detect the filesystem type of /system rather than explicitly specify EXT4, and changes it to instead of format /system as EXT4 to run rm -rf which accomplishes the same task.

This allows for flashing on both an EXT4-formatted /system and an F2FS-formatted /system without needing any 'patches' to the updater-script.

Change-Id: Ieec13451b30b6c821f5c6614c6baab1f6ef35315
Signed-off-by: Kyle Repinski <repinski23@gmail.com>