tuna: simplify /factory partition handling
Currently, we do the following sequence on every bootup:
1. mount /factory ro (fstab)
2. remount /factory as rw
3. update selinux labels on the partition
4. remount /factory as ro

By mounting /factory in rw mode from the fstab, we can avoid the second step,
simplifying the bootup sequence.

Also replaced the restorecons with one restorecon_recursive,
which does the same task in one line.

And lastly, removed some unnecessary chown/chmods - remounting
does not overwrite the permissions and the owner.